SaveChangeAsync Interceptor with ChangeTracker

+++ Core/ETicaretAPI.Domain/Entities/Product.cs
@@ -11,7 +11,7 @@ namespace ETicaretAPI.Domain.Entities
     {
         public string Name { get; set; }
         public int Stock { get; set; }
-        public long Price { get; set; }
+        public float Price { get; set; }
         public ICollection<Order> Orders { get; set; }
     }
 }
